clean up restore case
