+++ src/components/repay/RepayAnvil.tsx
@@ -235,9 +235,7 @@ export const RepayAnvil: React.FC<ReserveComponentProps> = ({ poolId, assetId })
             )}
           </TxOverview>
         )}
-        {isError && (
-          <AnvilAlert severity={disabledType} message={reason} simResponse={simResponse} />
-        )}
+        {isError && <AnvilAlert severity={disabledType} message={reason} />}
       </Section>
     </Row>
   );
