+++ src/components/withdraw/WithdrawAnvil.tsx
@@ -270,12 +270,7 @@ export const WithdrawAnvil: React.FC<ReserveComponentProps> = ({ poolId, assetId
           </TxOverview>
         )}
         {isError && (
-          <AnvilAlert
-            severity={disabledType}
-            message={reason}
-            extraContent={extraContent}
-            simResponse={simResponse}
-          />
+          <AnvilAlert severity={disabledType} message={reason} extraContent={extraContent} />
         )}
       </Section>
     </Row>
